# Cranelift: remove logging of vcode when the log level isn't debug or more