--- a/cranelift/codegen/src/machinst/compile.rs
+++ b/cranelift/codegen/src/machinst/compile.rs
@@ -5,7 +5,7 @@ use crate::machinst::*;
 use crate::settings;
 use crate::timing;
 
-use log::debug;
+use log::{debug, log_enabled, Level};
 use regalloc::{allocate_registers_with_opts, Algorithm, Options, PrettyPrint};
 
 /// Compile the given function down to VCode with allocated registers, ready
@@ -29,10 +29,15 @@ where
         lower.lower(b)?
     };
 
-    debug!(
-        "vcode from lowering: \n{}",
-        vcode.show_rru(Some(b.reg_universe()))
-    );
+    // Creating the vcode string representation may be costly for large functions, so don't do it
+    // if the Debug level hasn't been statically (through features) or dyanmically (through
+    // RUST_LOG) enabled.
+    if log_enabled!(Level::Debug) {
+        debug!(
+            "vcode from lowering: \n{}",
+            vcode.show_rru(Some(b.reg_universe()))
+        );
+    }
 
     // Perform register allocation.
     let (run_checker, algorithm) = match vcode.flags().regalloc() {
@@ -101,10 +106,12 @@ where
         vcode.replace_insns_from_regalloc(result);
     }
 
-    debug!(
-        "vcode after regalloc: final version:\n{}",
-        vcode.show_rru(Some(b.reg_universe()))
-    );
+    if log_enabled!(Level::Debug) {
+        debug!(
+            "vcode after regalloc: final version:\n{}",
+            vcode.show_rru(Some(b.reg_universe()))
+        );
+    }
 
     Ok(vcode)
 }