This logging step may be quite expensive, since rendering the VCode into strings has never been
optimized at all and is using pretty naive string concatenation. Removing it is a clear compile times win for a large wasm module I'm looking at. These are the results with parallel compilation enabled, so there could be an overall speedup due to less malloc lock contention too.

- x64 (AMD Ryzen 9 3950X): from 5.2 seconds to 1.8 seconds, i.e. 65% speedup
- aarch64 (Mac M1): from 17.6 seconds to 6.4 seconds, i.e. 63% speedup