@cfallin 2021-03-23
This is fantastic, thanks! I'm curious: what logging crate did you have in your environment (or was this just in wasmtime)? There was a related issue we found recently that was fixed in tokio-rs/tracing#1247; though if a similar issue occurs with e.g. `env_logger` then that's good to know too, and we should audit regalloc.rs for unnecessary expression evals in tracing!