@alexcrichton 2021-03-23
For posterity we had [more discussion on Zulip](https://bytecodealliance.zulipchat.com/#narrow/stream/217117-cranelift/topic/logging) about this and I believe that this is actually not necessary and I shouldn't have approved so quickly. I was under the impression that the arguments to logging functions were always evaluated but that is not the case. Further investigation shows that @bnjbvr's local logger may be behaving badly which was causing the slowdown here.