@bnjbvr 2021-03-23
Yes, the `log::max_level` was set to the highest possible value by a dependency's default ctor. Sorry for the churn, will post a revert soonish.